Remove utils module & move files/tests
