+++ fitting/greedy/greedy_lq.py
@@ -27,8 +27,8 @@ import numpy as np
 
 from fitting.greedy.greedy_strat import GreedyStrategy
 from fitting.greedy.greedy_utils import get_next_choices
+from fitting.greedy.optimize import optimize_using_nnls, optimize_using_slsqp
 from fitting.model import AtomicGaussianDensity
-from fitting.utils.optimize import optimize_using_nnls, optimize_using_slsqp
 
 __all__ = ["GreedyLeastSquares"]
