+++ fitting/greedy/optimize.py
